+++ src/network.go
@@ -37,9 +37,9 @@ type GetBuildResponse struct {
 }
 
 type RegisterRunnerRequest struct {
-	Token    string `json:"token,omitempty"`
-	Hostname string `json:"hostname,omitempty"`
-	Tags     string `json:"tag_list,omitempty"`
+	Token       string `json:"token,omitempty"`
+	Description string `json:"description,omitempty"`
+	Tags        string `json:"tag_list,omitempty"`
 }
 
 type RegisterRunnerResponse struct {
@@ -149,11 +149,11 @@ func GetBuild(config RunnerConfig) (*GetBuildResponse, bool) {
 	}
 }
 
-func RegisterRunner(url, token, hostname, tags string) *RegisterRunnerResponse {
+func RegisterRunner(url, token, description, tags string) *RegisterRunnerResponse {
 	request := RegisterRunnerRequest{
-		Token:    token,
-		Hostname: hostname,
-		Tags:     tags,
+		Token:       token,
+		Description: description,
+		Tags:        tags,
 	}
 
 	var response RegisterRunnerResponse