+++ CHANGELOG.md
@@ -1,2 +1,14 @@
+v 0.1.3
+- Optimize setup
+- Optimize multi-runner setup - making it more concurrent
+- Send description instead of hostname during registration
+- Don't ask for tags
+
+v 0.1.2
+- Make it work on Windows
+
+v 0.1.1
+- Added Docker services
+
 v 0.1.0
 - Initial public release